01-broken-access-control/abac/requirements.txt
# web-security-snippets — 01-broken-access-control/abac
# Install: pip install -r requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
Flask==3.0.2
PyJWT==2.8.0
pydantic==2.6.3
python-jose==3.3.0
